% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_data_help_gen.R
\docType{data}
\name{aduf}
\alias{aduf}
\title{Fake average daily active users per week over the course of a year}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 13 rows and 4 columns.
}
\usage{
data(aduf)
}
\description{
Fake (simulated) average daily active users per week over the course of a year on a fake
social media site. Each row of the data is a month. Each column represents one of the four
weeks in a month. Simulation based on insights found here: https://financesonline.com/number-of-twitter-users/
}
\examples{
data(aduf)
}
\references{
https://financesonline.com/number-of-twitter-users/
}
\keyword{datasets}
