% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_cox.R
\name{box.cox.transform}
\alias{box.cox.transform}
\title{Convience wrapper for Box-Cox transform}
\usage{
box.cox.transform(dat, lambda = NULL)
}
\arguments{
\item{dat}{a vector or array of data}

\item{lambda}{chosen box-cox lambda}
}
\value{
The function will XX
}
\description{
Convience wrapper for Box-Cox transform
}
\details{
If no lambda is given (NULL), no transformation is performed.
}
