% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{bs.plot.with.obs.statistic}
\alias{bs.plot.with.obs.statistic}
\title{Plot a bootstrapped sample with CIs and observed sample statistic.}
\usage{
bs.plot.with.obs.statistic(
  dat.bs,
  obs.statistic,
  confidence = 0.95,
  alternative = "two.sided",
  two.sided.ref = NULL,
  aslQ = F,
  xlim.marg = 0.1
)
}
\arguments{
\item{XX}{A bootstrap sample}
}
\value{
The function will XX
}
\description{
Plot a bootstrapped sample with CIs and observed sample statistic
}
