% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{bs.test}
\alias{bs.test}
\title{Non-Parametric Bootstrap Hypothesis Testing.}
\usage{
bs.test(
  x,
  y = NULL,
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  conf.level = 0.95,
  paired = FALSE,
  num.B = 2000,
  xlim.marg = 0.1
)
}
\arguments{
\item{x}{Sample of data to be non-parametrically bootstrapped}

\item{y}{Sample of data to be non-parametrically bootstrapped}
}
\value{
The function will XX
}
\description{
Non-Parametric Bootstrap Hypothesis Testing. Set up to have similar inputs, outputs and function to t.test()
}
