% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_data_help_gen.R
\docType{data}
\name{cms.km}
\alias{cms.km}
\title{Neel and Wells KM CMS data combined with Petraco group automatically extracted KM CMS data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 10 rows and 10 columns.
}
\usage{
data(cms.km)
}
\description{
Neel and Wells known match (KM) consecutive matching striae (CMS) data combined with Petraco group automatically extracted CMS data
}
\examples{
data(cms.km)
}
\references{
Neel, M and Wells M. “A Comprehensive Analysis of Striated Toolmark Examinations. Part 1: Comparing Known Matches to Known Non-Matches”, AFTE J 39(3):176-198 2007.

N. D. K. Petraco, L. Kuo, H. Chan, E. Phelps, C. Gambino, P. McLaughlin, Frani Kammerman, P. Diaczuk, P. Shenkin, N. Petraco and J. Hamby, "Estimates of Striation Pattern Identi cation Error Rates by Algorithmic Methods," AFTE J. 45(3), 235 (2013)
}
\keyword{datasets}
