% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_data_help_gen.R
\docType{data}
\name{dna_touch_drift}
\alias{dna_touch_drift}
\title{Fake drifting touch DNA data as a function of air temperature and room air flow}
\format{
An object of class \code{data.frame} with 111 rows and 3 columns.
}
\usage{
data(dna_touch_drift)
}
\description{
Fake DNA air flow data generated to look reasonable. This data has nice interactions and is useful
for illustrating regression models.
}
\examples{
data(dna_touch_drift)
}
\references{
The R Book
}
\keyword{datasets}
