% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_data_help_gen.R
\docType{data}
\name{dropout.info}
\alias{dropout.info}
\title{Fake allelic dropout data generated from Tvedebrink 2012, 28-cycle logistic regression coefficients.}
\format{
An object of class \code{list} of length 11.
}
\usage{
data(dropout.info)
}
\description{
Fake allelic dropout data generated from Tvedebrink 2012, 28-cycle logistic regression coefficients.
Data is a list of matrices for loci: "D3", "vWA", "D16", "D2", "D8", "SE33", "D19", "TH0", "FGA",
"D21" and "D18". Columns of the matrices are H (related to alleal peak height(s) at each locus),
log(H) and a dropout indicator (D.vec). In the D.vec column, 1 = peak dropped out, 0 = peak present.
D.vec data is fake and simulated using Tvedebrink 2012 b0 and b1 coefficients. Simulation model:
b0 ~ normal(b0-2012, b0se-2012), b1 ~ normal(b1-2012, b1se-2012), log-Odds = b0 + b1*logH,
pD = p(dropout) = inv-logit(pD) and D.vec[i] = D.vec(H) ~ bernoulli(pD(H)). Logistic regression
coefficients used in the model are contained in tvedebrink12.
}
\examples{
data(dropout.info)
}
\references{
T Tvedebrink, PS Eriksen, M Asplund, HS Mogensen, N Morling. Allelic drop-out
probabilities estimated by logistic regression—Further considerations and practical implementation.
Forensic Science International: Genetics 6 (2012) 263–267.
}
\keyword{datasets}
