% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_data_help_gen.R
\docType{data}
\name{frisk}
\alias{frisk}
\title{Gelman et al. stop and frisk data with noise added to protect confidentiality}
\format{
An object of class \code{data.frame} with 900 rows and 6 columns.
}
\usage{
data(frisk)
}
\description{
Gelman et al. stop and frisk data with noise added to protect confidentiality. Precincts are
numbered 1-75, ethnicity 1=black, 2=hispanic, 3=white, crime type 1=violent, 2=weapons, 3=property,
4=drug.
}
\examples{
data(frisk)
}
\references{
A Gelman, J Fagan, A Kiss. An Analysis of the New York City Police Department’s
"Stop-and-Frisk" Policy in the Context of Claims of Racial Bias. Journal of the American Statistical
 Association, Vol. 102, No. 479, Pp. 813-823, 2007.

http://www.stat.columbia.edu/~gelman/arm/examples/police/frisk_with_noise.dat
}
\keyword{datasets}
