% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_data_help_gen.R
\docType{data}
\name{gas}
\alias{gas}
\title{Mark Gil et al. gasoline data}
\format{
An object of class \code{data.frame} with 92 rows and 16 columns.
}
\usage{
data(gas)
}
\description{
Data collected by Mark Gil et al. on the content of 15 hydrocarbons found in gasolines
}
\examples{
data(gas)
}
\references{
NDK Petraco, M Gil, PA Pizzola, TA Kubic, "Statistical Discrimination of Liquid Gasoline Samples from Casework", J Forensic Sci 53(5), 1092-1101 (2008)
}
\keyword{datasets}
