% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{get.script}
\alias{get.script}
\title{Download and load a script file}
\usage{
get.script(script.name, url_head = NULL, download_location = NULL)
}
\arguments{
\item{XX}{XX}
}
\value{
The function will XX
}
\description{
Download and load a script file. Intended to get script files for the Notes
}
