% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{grubbs.parametric.bs.test}
\alias{grubbs.parametric.bs.test}
\title{Grubbs outlier test as a parametric bootstrap.}
\usage{
grubbs.parametric.bs.test(
  xdat,
  rdist.func,
  rdist.args.list,
  confidence.level = 0.95,
  num.B = 2000
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Grubbs outlier test as a parametric bootstrap.
}
