% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_data_help_gen.R
\docType{data}
\name{indkm.counts}
\alias{indkm.counts}
\title{Simulated counts of corresponding minutae for mated pairs (KM) of latent fingerprints}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1801 rows and 1 columns.
}
\usage{
data(indkm.counts)
}
\description{
Simulated examiner designated counts of corresponding minutae for mated pairs (KM) of
latent fingerprints. Simulations based on Ulrey 2014 with 165 examiners judging 231
mated pairs. There are 1801 examinations of mated pairs among these simulated
examiners. Each entry represents how many corresponding minutae were designated
leading the examiner to judge the pair a match. Summary statistics of this dataset
are consistent with the Ulrey 2014 paper.
}
\examples{
data(indkm.counts)
}
\references{
BT Ulery, RA Hicklin, MA Roberts, J Buscaglia. Measuring What Latent
Fingerprint Examiners Consider Sufficient Information for Individualization
Determinations; PLoS ONE 2014; 9(11): e110179. https://doi.org/10.1371/journal.pone.0110179
}
\keyword{datasets}
