% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_data_help_gen.R
\docType{data}
\name{nba}
\alias{nba}
\title{NBA player data}
\format{
An object of class \code{data.frame} with 54 rows and 5 columns.
}
\usage{
data(nba)
}
\description{
Data for some anonymous NBA players.
ht = height,
wt = weight,
fg.perc = field goal percentage,
ft.perc = free throw percentage,
avg.ppg = average points per game
}
\examples{
data(nba)
}
\references{
The official NBA basketball Encyclopedia, Villard Books, 1989.
}
\keyword{datasets}
