% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{one.sample.bs.hyp.for.mean}
\alias{one.sample.bs.hyp.for.mean}
\title{One sample Non-Parametric Bootstrap Hypothesis Test for a Mean.}
\usage{
one.sample.bs.hyp.for.mean(
  dat.samp,
  mu.null,
  confidence = 0.95,
  alternative = "two.sided",
  num.B = 2000,
  xlim.marg = 0.1
)
}
\arguments{
\item{dat.samp}{Sample of data to be non-parametrically bootstrapped}

\item{mu.null}{Fixed (null) value for mean. Compare sample mean to this.}
}
\value{
The function will XX
}
\description{
One sample Non-Parametric Bootstrap Hypothesis Test for a Mean.
}
