% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing_BROKEN.R
\name{one.sample.bs.hyp2_BROKEN}
\alias{one.sample.bs.hyp2_BROKEN}
\title{One sample Non-Parametric Bootstrap Hypothesis Tests with Studentizing option.}
\usage{
one.sample.bs.hyp2_BROKEN(
  dat.samp,
  stat.func = mean,
  val.null,
  studentizeQ = F,
  confidence = 0.95,
  alternative = "two.sided",
  num.B = 2000
)
}
\arguments{
\item{dat.samp}{Sample of data to be non-parametrically bootstrapped}

\item{stat.func}{Statistical function to bootstrap. Usually a mean, but can be anything that take a single vector of data and produces a statistic.}

\item{val.null}{Fixed (null) value to compare sample statistic to. Usually a stipulated reference mean.}

\item{studentizeQ}{Whether or not to Studentize the test statistic}
}
\value{
The function will XX
}
\description{
One sample Non-Parametric Bootstrap Hypothesis Tests Studentizing option.
}
