% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{pairwise.tests}
\alias{pairwise.tests}
\title{Convenience wrapper for pairwise t-tests and Wilcoxon tests}
\usage{
pairwise.tests(
  xvec,
  glbl,
  confidence = 0.95,
  significant.onlyQ = F,
  type = c("t.test", "wilcox.test"),
  p.adjust.method,
  pool.sd = !paired,
  paired = FALSE,
  alternative = "two.sided"
)
}
\arguments{
\item{XX}{XXXX}
}
\value{
The function will XX
}
\description{
Convenience wrapper for pairwise t-tests and Wilcoxon tests
}
