% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{percentiles}
\alias{percentiles}
\title{Wraper function to give percentiles of input data with respect to an empirical distribution}
\usage{
percentiles(all.dat, quantiles)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Wraper function to give percentiles of input data with respect to an empirical distribution
}
