% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_data_help_gen.R
\docType{data}
\name{radon}
\alias{radon}
\title{Gelman and Hill data on radon levels in houses in the state of Minnesota.}
\format{
An object of class \code{data.frame} with 919 rows and 4 columns.
}
\usage{
data(radon)
}
\description{
Data taken directly from rstanarm
919 obs. of 4 variables
log_radon Radon measurement from the house (log scale)
log_uranium Uranium level in the county (log scale)
floor Indicator for radon measurement made on the first floor of the house (0 = basement, 1 = first floor)
county County name (factor)
}
\examples{
data(nba)
}
\references{
Gelman and Hill, Cambridge University Press, 2006.
}
\keyword{datasets}
