% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{rule.of.thumb.outliers}
\alias{rule.of.thumb.outliers}
\title{Gage data in a data set as outliers using the standard rules of thumb.}
\usage{
rule.of.thumb.outliers(xdat, plotQ = F)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Gage data in a data set as outliers using the standard rules of thumb.
}
