% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_data_help_gen.R
\docType{data}
\name{sgr.nsim}
\alias{sgr.nsim}
\title{Fake data generated from from shotgun.df in Curran's dafs}
\format{
An object of class \code{data.frame} with 60 rows and 6 columns.
}
\usage{
data(sgr.nsim)
}
\description{
Simulated square-root area shotgun pellet pattern from data collected by Rowe and Hanson (contained
in James Curran's dafs package). The simulation uses a Bayesian model to obtain posterior distributions
for each distance's mean and sd. Fake data is then obtained using these posterior parameter
distributions with a normal distribution (hence .nsim) to simulate data. 60 fake data points are then
sampled from the simulated dataset (30 Stevens, 30 Remington).
}
\examples{
data(sgr.nsim)
}
\references{
Rowe, W.F. and Hanson, S.R. "Range-of-fire estimates from regression analysis applied to the spreads of shotgun pellet patterns: Results of a blind study", Forensic Sci Int, 28(3-4):239-250, (1985).
}
\keyword{datasets}
