% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_data_help_gen.R
\docType{data}
\name{sgr}
\alias{sgr}
\title{Experimental data from shotgun.df in Curran's dafs}
\format{
An object of class \code{data.frame} with 12 rows and 6 columns.
}
\usage{
data(sqr)
}
\description{
Reformatted square-root area shotgun pellet pattern data collected by Rowe and Hanson.
Contained in James Curran's dafs package. Experimental (train) data only, cf. data(shotgun.df) in dafs.
}
\examples{
data(sgr)
}
\references{
Rowe, W.F. and Hanson, S.R. "Range-of-fire estimates from regression analysis applied to the spreads of shotgun pellet patterns: Results of a blind study", Forensic Sci Int, 28(3-4):239-250, (1985).
}
\keyword{datasets}
