% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_data_help_gen.R
\docType{data}
\name{sgr.tsim}
\alias{sgr.tsim}
\title{More fake data generated from from shotgun.df in Curran's dafs}
\format{
An object of class \code{data.frame} with 60 rows and 6 columns.
}
\usage{
data(sgr.tsim)
}
\description{
Simulated square-root area shotgun pellet pattern from data collected by Rowe and Hanson (contained
in James Curran's dafs package). The simulation uses a Bayesian model to obtain posterior distributions
for each distance's mean and sd. Fake data is then obtained using these posterior parameter
distributions with a student-T distribution (hence .tsim) to simulate data. 60 fake data points are then
sampled from the simulated dataset (30 Stevens, 30 Remington). Using the student-T as the data's
distribution also adds in more uncertainty. The (fake) data is also drawn from the Bayeisan model's
posterior predictive distribution in this case (unlike in sgr.nsim).
}
\examples{
data(sgr.tsim)
}
\references{
Rowe, W.F. and Hanson, S.R. "Range-of-fire estimates from regression analysis applied to the spreads of shotgun pellet patterns: Results of a blind study", Forensic Sci Int, 28(3-4):239-250, (1985).
}
\keyword{datasets}
