% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_data_help_gen.R
\name{shedder.donor.boxplot}
\alias{shedder.donor.boxplot}
\title{Utility to help plot simplified (fake) shedder donor data}
\usage{
shedder.donor.boxplot(
  shedder.data,
  location.variable,
  response.variable,
  response.statistic,
  orderQ = T
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Utility to help plot simplified (fake) shedder donor data
}
\details{
The complete "shedder" data is complicated. Analyzing it all at once is not easy. The function is
to plot the "shedder" data as a function of the donors for visualization purposes.
}
\examples{
AN EXAMPLE HERE

}
