% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_data_help_gen.R
\name{shedder.extract.data}
\alias{shedder.extract.data}
\title{Utility to help simplify (fake) shedder data}
\usage{
shedder.extract.data(
  shedder.data,
  location.variable,
  response.variable = "amount.DNA",
  response.statistic,
  orderQ = F
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Utility to help simplify (fake) shedder data
}
\details{
The complete "shedder" data is complicated. Analyzing it all at once is not easy. The function is
to help slice out relevant parts of the  "shedder" data for visualization and analysis purposes.
}
