% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_data_help_gen.R
\docType{data}
\name{shedder}
\alias{shedder}
\title{Fake DNA shedder data}
\format{
An object of class \code{data.frame} with 350 rows and 5 columns.
}
\usage{
data(shedder)
}
\description{
Fake (simulated) shed amounts of touch DNA. Simulation based on MS thesis work of Xiao Chen.
(Fake) Shed amounts of touch DNA are for 50 (fake) participants with three replicate assays each
from seven locations on the body. A "W" suffix indicates washed, a "U" suffix indicates unwashed.
}
\examples{
data(shedder)
}
\references{
X. Chen "Evaluating a Test for Shedding Propensity Using Tape Lifts from Different Skin Locations", Dec-27-2021.

https://academicworks.cuny.edu/jj_etds/199/
}
\keyword{datasets}
