% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_data_help_gen.R
\name{shedder.summary.extract.data}
\alias{shedder.summary.extract.data}
\title{Utility to help summarize simplified (fake) shedder data}
\usage{
shedder.summary.extract.data(
  shedder.extracted.data,
  response.variable,
  order.statistic = mean,
  orderQ = T
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Utility to help summarize simplified (fake) shedder data
}
\details{
The complete "shedder" data is complicated. Analyzing it all at once is not easy. The function is
to help slice out relevant parts of the "shedder" data and summarize it for visualization and analysis purposes.
}
