% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{tm.parametric.bs.test}
\alias{tm.parametric.bs.test}
\title{Tietjen-Moore outlier test as a parametric bootstrap.}
\usage{
tm.parametric.bs.test(
  xdat,
  k,
  standardizeQ = T,
  rdist.func = rnorm,
  rdist.args.list = list(n = length(xdat), mean = 0, sd = 1),
  confidence.level = 0.95,
  num.B = 2000
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Tietjen-Moore outlier test as a parametric bootstrap.
}
\details{
Set up to run assuming data is normal and should be standardized
}
