% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{tmts}
\alias{tmts}
\title{Tietjen-Moore two-sided test statistic. From NIST Handbook code.}
\usage{
tmts(xx, kk)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Tietjen-Moore two-sided test statistic. From NIST Handbook code.
}
