% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_data_help_gen.R
\docType{data}
\name{transactions}
\alias{transactions}
\title{Fake financial fraud transaction data}
\format{
An object of class \code{data.frame} with 60 rows and 2 columns.
}
\usage{
data(transactions)
}
\description{
A simple faked financial transaction data set consisting of simulated suspicion scores for a group
of 60 transactions. 30 transactions were "ruled" not to be involved in fraud and 30 transactions
were "ruled" to be involved in fraud.
}
\examples{
data(transactions)
}
\references{
RJ Bolton, DJ Hand. "Statistical Fraud Detection: A review", Statist. Sci. 17(3): 235-255 (2002)

https://projecteuclid.org/journalArticle/Download?urlId=10.1214%2Fss%2F1042727940
}
\keyword{datasets}
