% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{two.sample.bs.hyp.for.mean}
\alias{two.sample.bs.hyp.for.mean}
\title{Two sample Non-Parametric Bootstrap Hypothesis Test to test difference between means.}
\usage{
two.sample.bs.hyp.for.mean(
  A.samp,
  B.samp,
  confidence = 0.95,
  alternative = "two.sided",
  num.B = 2000,
  xlim.marg = 0.1,
  print.type = 1
)
}
\arguments{
\item{A.samp}{Sample of data to be non-parametrically bootstrapped}

\item{B.samp}{Sample of data to be non-parametrically bootstrapped}
}
\value{
The function will XX
}
\description{
Two sample Non-Parametric Bootstrap Hypothesis Test to test difference between means.
}
