% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_frequtils}
\alias{update_frequtils}
\title{Easily update the frequtils library by installing the current version from the github site}
\usage{
update_frequtils()
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Easily update the frequtils library by installing the current version from the github site
}
