% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_data_help_gen.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine quality/chemical data}
\format{
An object of class \code{data.frame} with 6497 rows and 13 columns.
}
\usage{
data(wine)
}
\description{
Data for red and white variants of the Portuguese "Vinho Verde" wine and associated quality judgement scores.
}
\examples{
data(wine)
}
\references{
P. Cortez, A. Cerdeira, F. Almeida, T. Matos and J. Reis. Modeling wine preferences by data mining from physicochemical properties. In Decision Support Systems, Elsevier, 47(4):547-553, 2009.
}
\keyword{datasets}
