% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_data_help_gen.R
\docType{data}
\name{wine2}
\alias{wine2}
\title{Wine cultivar data}
\format{
An object of class \code{data.frame} with 178 rows and 14 columns.
}
\usage{
data(wine2)
}
\description{
These data are the results of a chemical analysis of wines grown in the same region in Italy but derived from three different cultivars. The analysis determined the quantities of 13 constituents found in each of the three types of wines.
This is an old dataset and not sure of the original reference. See the link to the UCI database below.
}
\examples{
data(wine2)
}
\references{
1. cf. https://archive.ics.uci.edu/ml/datasets/Wine

2. Forina, M. et al, PARVUS -An Extendible Package for Data Exploration, Classification and Correlation. Institute of Pharmaceutical and Food Analysis and Technologies, Via Brigata Salerno, 16147 Genoa, Italy.

3. UCI Donor: Stefan Aeberhard, email: stefan-at-coral.cs.jcu.edu.au
}
\keyword{datasets}
