% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_cox.R
\name{box.cox.transform}
\alias{box.cox.transform}
\title{Convience wrapper for Box-Cox transform}
\usage{
box.cox.transform(dat, lambda)
}
\arguments{
\item{XX}{XX}
}
\value{
The function will XX
}
\description{
Convience wrapper for Box-Cox transform
}
